Simplify the test app
